requirements.txt
APScheduler==3.8.1
coreapi<=2.3.3
cryptography<=37.0.2
Django <=3.0.14
djangorestframework == 3.10.3
django-import-export<=2.7.1
django-admin-menu<=1.5 #checkout stable version
data-wizard==2.0.0
git+https://github.com/de-laz/openhim-mediator-utils-py
git+https://github.com/ahoazure/aho_dhis2dct_taskscheduler.git
mysqlclient<=2.0.1
python-dateutil <=2.8.2
python-dotenv<=0.17.1
requests >= 2.18.4,<=2.27.1
wheel
whitenoise <= 5.3.0

# Data science and world bank packages
pandas==1.4.3
numpy==1.23.2
django-pandas==0.6.6
wbgapi<=1.0.12 